+++ CHANGES.md
@@ -34,6 +34,7 @@ Update version in extension.json.
 YYYY-MM-DD
 Speechoid: 0.1.2
 
+* [T280244](https://phabricator.wikimedia.org/T280244) Optionally mark utterance files with TTL in Swift.
 * [T133857](https://phabricator.wikimedia.org/T133857) Only add edit button to the player if user has right to edit lexicon
 * [T280239](https://phabricator.wikimedia.org/T280239) Use GET requests for segments and utterances